refactor(blackjack/service): 서비스 레이어 삭제
불필요한 서비스 레이어 삭제
