+++ src/main/java/blackjack/BlackJackApplication.java
@@ -1,7 +1,5 @@
 package blackjack;
 
-import blackjack.service.BlackJack;
-
 public class BlackJackApplication {
 
 	public static void main(String[] args) {
