+++ src/main/java/blackjack/domain/card/Hand.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import blackjack.service.BlackJack;
+import blackjack.BlackJack;
 
 public class Hand {
